testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/combination_history_001.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>combination_history_001(Combine pushState and replaceSate methods.)</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		try {
			var first;
			var second;
			window.history.pushState(1, document.title, '?x=1');
			first = window.history.state;
			window.history.replaceState(2, document.title, '?x=1');
			second = window.history.state;

			if (first == 1 && second == 2) {
				test(function() {
					assert_true(true, "combination test is failed.");
				});
			} else {
				test(function() {
					assert_unreached("combination test is failed.");
				});
			}
		} catch (e) {
			test(function() {
				assert_unreached("call of pushState method is failed.");
			});
		}

	</script>
 </body>
</html>
